actions and reducers now organised by modules
