--- src/js/actions/actionTypes.js
@@ -1,30 +1,0 @@
-module.exports = {
-  SET_VIEW: 'SET_VIEW',
-  STEP_INTRO_TEXT: 'STEP_INTRO_TEXT',
-  RESET_INTRO_STEP: 'RESET_INTRO_STEP',
-
-  SET_DRAG: 'SET_DRAG',
-  START_DRAG: 'START_DRAG',
-  STOP_DRAG: 'STOP_DRAG',
-
-  IS_UPDATING: 'IS_UPDATING',
-  CHECK_TILE: 'CHECK_TILE',
-  SET_ENTERING: 'SET_ENTERING',
-
-  UPDATE_SCORE: 'UPDATE_SCORE',
-  WEATHER_POWER: 'WEATHER_POWER',
-  RESET_WEATHER: 'RESET_WEATHER',
-
-  GROW_SEEDS: 'GROW_SEEDS',
-  FALL_TILES: 'FALL_TILES',
-  SHIFT_TILES: 'SHIFT_TILES',
-  TRANSFORM_BOARD: 'TRANSFORM_BOARD',
-  ADD_TILES: 'ADD_TILES',
-
-  REMOVE_SEEDS: 'REMOVE_SEEDS',
-
-  RESET_GROW_SEEDS: 'RESET_GROW_SEEDS',
-  RESET_LEAVING: 'RESET_LEAVING',
-  RESET_MAGNITUDE: 'RESET_MAGNITUDE',
-  RESET_ENTERING: 'RESET_ENTERING'
-}
