+++ src/js/model/model.js
@@ -56,7 +56,7 @@ const mapWithIndex = addIndex(map)
 
 // converts the tiles on the board to a new number
 // where the board coordiantes match the coordiantes in an array of moves
-export const transformBoard = (moves, board, transformNumber) =>
+export const transformTiles = (moves, board, transformNumber) =>
   mapWithIndex((row, i) =>
   mapWithIndex((tile, j) =>
   filter(([ y, x ]) =>
