--- src/js/redux/modules/reducer_setDrag.js
@@ -1,10 +1,0 @@
-import { SET_DRAG } from '../../actions/actionTypes.js'
-
-export default (state = false, action) => {
-  switch (action.type) {
-  case SET_DRAG:
-    return action.payload
-  default:
-    return state
-  }
-}
