+++ src/js/redux/root_reducer.js
@@ -1,10 +1,10 @@
 import { combineReducers } from 'redux'
-import text from '../modules/reducer_text.js'
-import isDragging from '../modules/reducer_setDrag.js'
+import text from '../modules/text.js'
+import isDragging from '../modules/setDrag.js'
 import updating from '../modules/isUpdating.js'
-import weather from '../modules/reducer_weatherPower.js'
-import score from '../modules/reducer_score.js'
-import view from '../modules/reducer_view.js'
+import weather from '../modules/weatherPower.js'
+import score from '../modules/score.js'
+import view from '../modules/view.js'
 import moves from '../modules/moves.js'
 import board from '../modules/board.js'
 import fallingMagnitudeArray from '../modules/fallingMagnitude.js'
